requirements.txt
aiogram[payments]
python-dotenv
openai
aiosqlite